contranss-api/requirements.txt
Django==1.7.8
MySQL-python==1.2.5
PyJWT==1.3.0
PyYAML==3.11
argparse==1.2.1
defusedxml==0.4.1
diff-match-patch==20121119
django-filter==0.10.0
django-import-export==0.2.7
django-rest-swagger==0.3.2
djangorestframework==3.1.3
djangorestframework-filter==0.2.2
djangorestframework-filters==0.3.0
djangorestframework-jwt==1.6.0
djangorestframework-xml==1.2.0
elasticsearch==1.6.0
tablib==0.10.0
unicodecsv==0.13.0
urllib3==1.10.4
wsgiref==0.1.2
